+++ src/Entity/Sites.php
@@ -2,6 +2,8 @@
 
 namespace App\Entity;
 
+use Doctrine\Common\Collections\ArrayCollection;
+use Doctrine\Common\Collections\Collection;
 use Doctrine\ORM\Mapping as ORM;
 
 /**
@@ -28,6 +30,7 @@ class Sites
      */
     private $nomSite;
 
+
     public function getNoSite(): ?int
     {
         return $this->noSite;
@@ -44,6 +47,7 @@ class Sites
 
         return $this;
     }
+
 
 
 }
