--- src/Entity/User.php
@@ -1,185 +1,0 @@
-<?php
-
-namespace App\Entity;
-
-use App\Repository\UserRepository;
-use Symfony\Component\Validator\Constraints as Assert;
-use Doctrine\ORM\Mapping as ORM;
-use Symfony\Bridge\Doctrine\Validator\Constraints\UniqueEntity;
-use Symfony\Component\Security\Core\User\PasswordAuthenticatedUserInterface;
-use Symfony\Component\Security\Core\User\UserInterface;
-
-/**
- * @ORM\Entity(repositoryClass=UserRepository::class)
- * @UniqueEntity(fields={"pseudo"}, message="There is already an account with this pseudo")
- */
-class User implements UserInterface, PasswordAuthenticatedUserInterface
-{
-    /**
-     * @ORM\Id
-     * @ORM\GeneratedValue
-     * @ORM\Column(type="integer")
-     */
-    private $id;
-
-    /**
-     * @ORM\Column(type="string", length=180, unique=true)
-     */
-    private $pseudo;
-
-    /**
-     * @ORM\Column(type="json")
-     */
-    private $roles = [];
-
-    /**
-     * @var string The hashed password
-     * @ORM\Column(type="string")
-     */
-    private $password;
-
-    /**
-     * @ORM\OneToOne(targetEntity=Participants::class, mappedBy="userNoUser", cascade={"persist", "remove"})
-     */
-    private $participantId;
-
-     /**
-     * @Assert\Type(type="App\Entity\Participants")
-     * @Assert\Valid
-     */
-    protected $participant;
-
-    public function getId(): ?int
-    {
-        return $this->id;
-    }
-
-    public function getPseudo(): ?string
-    {
-        return $this->pseudo;
-    }
-
-    public function setPseudo(string $pseudo): self
-    {
-        $this->pseudo = $pseudo;
-
-        return $this;
-    }
-
-    /**
-     * A visual identifier that represents this user.
-     *
-     * @see UserInterface
-     */
-    public function getUserIdentifier(): string
-    {
-        return (string) $this->pseudo;
-    }
-
-    /**
-     * @deprecated since Symfony 5.3, use getUserIdentifier instead
-     */
-    public function getUsername(): string
-    {
-        return (string) $this->pseudo;
-    }
-
-    /**
-     * @see UserInterface
-     */
-    public function getRoles(): array
-    {
-        $roles = $this->roles;
-        // guarantee every user at least has ROLE_USER
-        $roles[] = 'ROLE_USER';
-
-        return array_unique($roles);
-    }
-
-    public function setRoles(array $roles): self
-    {
-        $this->roles = $roles;
-
-        return $this;
-    }
-
-    /**
-     * @see PasswordAuthenticatedUserInterface
-     */
-    public function getPassword(): string
-    {
-        return $this->password;
-    }
-
-    public function setPassword(string $password): self
-    {
-        $this->password = $password;
-
-        return $this;
-    }
-
-    /**
-     * Returning a salt is only needed, if you are not using a modern
-     * hashing algorithm (e.g. bcrypt or sodium) in your security.yaml.
-     *
-     * @see UserInterface
-     */
-    public function getSalt(): ?string
-    {
-        return null;
-    }
-
-    /**
-     * @see UserInterface
-     */
-    public function eraseCredentials()
-    {
-        // If you store any temporary, sensitive data on the user, clear it here
-        // $this->plainPassword = null;
-    }
-
-    public function getParticipants(): ?Participants
-    {
-        return $this->participants;
-    }
-
-    public function setParticipants(Participants $participants): self
-    {
-        // set the owning side of the relation if necessary
-        if ($participants->getUserNoUser() !== $this) {
-            $participants->setUserNoUser($this);
-        }
-
-        $this->participants = $participants;
-
-        return $this;
-    }
-
-    public function getParticipantId(): ?Participants
-    {
-        return $this->participantId;
-    }
-
-    public function setParticipantId(Participants $participantId): self
-    {
-        // set the owning side of the relation if necessary
-        if ($participantId->getUserNoUser() !== $this) {
-            $participantId->setUserNoUser($this);
-        }
-
-        $this->participantId = $participantId;
-
-        return $this;
-    }
-
-    public function getParticipant(): ?Participants
-    {
-        return $this->participant;
-    }
-
-    public function setCategory(?Participants $participant)
-    {
-        $this->participant = $participant;
-    }
-
-}
