--- src/Form/ParticipantsType.php
@@ -1,30 +1,0 @@
-<?php
-
-namespace App\Form;
-
-use App\Entity\Participants;
-use App\Entity\Sites;
-use Symfony\Bridge\Doctrine\Form\Type\EntityType;
-use Symfony\Component\Form\AbstractType;
-use Symfony\Component\Form\FormBuilderInterface;
-use Symfony\Component\OptionsResolver\OptionsResolver;
-
-class ParticipantsType extends AbstractType
-{
-    public function buildForm(FormBuilderInterface $builder, array $options): void
-    {
-        $builder
-            ->add('nom')
-            ->add('prenom')
-            ->add('telephone')
-            ->add('mail')
-        ;
-    }
-
-    public function configureOptions(OptionsResolver $resolver): void
-    {
-        $resolver->setDefaults([
-            'data_class' => Participants::class,
-        ]);
-    }
-}
