--- templates/participants/index.html.twig
@@ -1,45 +1,0 @@
-{% extends 'layouts/base.html.twig' %}
-
-{% block title %}Participants index{% endblock %}
-
-{% block main %}
-    <h1>Participants index</h1>
-
-    <table class="table">
-        <thead>
-            <tr>
-                <th>NoParticipant</th>
-                <th>Nom</th>
-                <th>Prenom</th>
-                <th>Telephone</th>
-                <th>Mail</th>
-                <th>Administrateur</th>
-                <th>Actif</th>
-                <th>actions</th>
-            </tr>
-        </thead>
-        <tbody>
-        {% for participant in participants %}
-            <tr>
-                <td>{{ participant.noParticipant }}</td>
-                <td>{{ participant.nom }}</td>
-                <td>{{ participant.prenom }}</td>
-                <td>{{ participant.telephone }}</td>
-                <td>{{ participant.mail }}</td>
-                <td>{{ participant.administrateur ? 'Yes' : 'No' }}</td>
-                <td>{{ participant.actif ? 'Yes' : 'No' }}</td>
-                <td>
-                    <a href="{{ path('app_participants_show', {'noParticipant': participant.noParticipant}) }}">show</a>
-                    <a href="{{ path('app_participants_edit', {'noParticipant': participant.noParticipant}) }}">edit</a>
-                </td>
-            </tr>
-        {% else %}
-            <tr>
-                <td colspan="10">no records found</td>
-            </tr>
-        {% endfor %}
-        </tbody>
-    </table>
-
-    <a href="{{ path('app_participants_new') }}">Create new</a>
-{% endblock %}
