--- templates/participants/new.html.twig
@@ -1,11 +1,0 @@
-{% extends 'layouts/base.html.twig' %}
-
-{% block title %}New Participants{% endblock %}
-
-{% block main %}
-    <h1>Create new Participants</h1>
-
-    {{ include('participants/_form.html.twig') }}
-
-    <a href="{{ path('app_participants_index') }}">back to list</a>
-{% endblock %}
