--- templates/participants/show.html.twig
@@ -1,46 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block title %}Participants{% endblock %}
-
-{% block body %}
-    <h1>Participants</h1>
-
-    <table class="table">
-        <tbody>
-            <tr>
-                <th>NoParticipant</th>
-                <td>{{ participant.noParticipant }}</td>
-            </tr>
-            <tr>
-                <th>Nom</th>
-                <td>{{ participant.nom }}</td>
-            </tr>
-            <tr>
-                <th>Prenom</th>
-                <td>{{ participant.prenom }}</td>
-            </tr>
-            <tr>
-                <th>Telephone</th>
-                <td>{{ participant.telephone }}</td>
-            </tr>
-            <tr>
-                <th>Mail</th>
-                <td>{{ participant.mail }}</td>
-            </tr>
-            <tr>
-                <th>Administrateur</th>
-                <td>{{ participant.administrateur ? 'Yes' : 'No' }}</td>
-            </tr>
-            <tr>
-                <th>Actif</th>
-                <td>{{ participant.actif ? 'Yes' : 'No' }}</td>
-            </tr>
-        </tbody>
-    </table>
-
-    <a href="{{ path('app_participants_index') }}">back to list</a>
-
-    <a href="{{ path('app_participants_edit', {'noParticipant': participant.noParticipant}) }}">edit</a>
-
-    {{ include('participants/_delete_form.html.twig') }}
-{% endblock %}
